+++ incremental/src/main/java/net/automatalib/incremental/dfa/dag/AbstractIncrementalDFADAGBuilder.java
@@ -297,6 +297,8 @@ public abstract class AbstractIncrementalDFADAGBuilder<I> extends AbstractIncrem
         register.remove(sig);
         sig.successors.array[idx] = succ;
         sig.acceptance = acc;
+        succ.increaseIncoming();
+        sig.updateHashCode();
         return replaceOrRegister(state);
     }